Tests/LibWeb/Text/expected/wheel-events-consumed-by-scrollable-should-not-be-propagated-to-body-2.txt

scrollable.scrollTop: 200
window.scrollY: 0
